threat-detection/requirements.txt
requests==2.26.0
tldextract==3.1.0
python-whois==0.8.0
scikit-learn==1.0.2
pandas==1.3.5